infra/rds/requirements.txt
git+https://github.com/kjenney/managor.git
mysql-connector-python>=8.0.22,<9.0.0
pulumi>=3.0.0,<4.0.0
pulumi-aws>=4.0.0,<5.0.0
